# Improve the performance of core PropertyPath code.