--- a/editor/src/components/canvas/controls/reorder-slider-control.tsx
+++ b/editor/src/components/canvas/controls/reorder-slider-control.tsx
@@ -11,7 +11,7 @@ import {
   zeroCanvasRect,
   zeroRectIfNullOrInfinity,
 } from '../../../core/shared/math-utils'
-import { arrayEquals } from '../../../core/shared/utils'
+import { arrayEqualsByValue } from '../../../core/shared/utils'
 import { Modifier } from '../../../utils/modifiers'
 import { when } from '../../../utils/react-conditionals'
 import { Icons, useColorTheme } from '../../../uuiui'
@@ -107,7 +107,7 @@ export const ReorderSliderControl = controlForStrategyMemoized(
       },
       'ReorderSliderControl',
       (oldValue, newValue) =>
-        arrayEquals(oldValue.siblings, newValue.siblings, EP.pathsEqual) &&
+        arrayEqualsByValue(oldValue.siblings, newValue.siblings, EP.pathsEqual) &&
         oldValue.latestIndex === newValue.latestIndex &&
         oldValue.startingIndex === newValue.startingIndex &&
         rectanglesEqual(oldValue.startingFrame, newValue.startingFrame),