--- a/editor/src/components/inspector/common/longhand-shorthand-hooks.ts
+++ b/editor/src/components/inspector/common/longhand-shorthand-hooks.ts
@@ -4,7 +4,7 @@ import { flatMapArray, last, mapArrayToDictionary } from '../../../core/shared/a
 import { emptyComments, jsExpressionValue } from '../../../core/shared/element-template'
 import { objectMap } from '../../../core/shared/object-utils'
 import { ElementPath } from '../../../core/shared/project-file-types'
-import { arrayEquals, NO_OP } from '../../../core/shared/utils'
+import { arrayEqualsByReference, NO_OP } from '../../../core/shared/utils'
 import { useKeepReferenceEqualityIfPossible } from '../../../utils/react-performance'
 import {
   setProp_UNSAFE,
@@ -43,7 +43,7 @@ function getShadowedLonghandShorthandValue<
   orderedPropKeys: (LonghandKey | ShorthandKey)[][], // multiselect
 ): { value: ParsedProperties[LonghandKey]; propertyStatus: PropertyStatus } {
   const allPropKeysEqual = orderedPropKeys.every((propKeys) => {
-    return arrayEquals(propKeys, orderedPropKeys[0])
+    return arrayEqualsByReference(propKeys, orderedPropKeys[0])
   })
 
   const propKeys = orderedPropKeys[0] ?? []
@@ -157,7 +157,7 @@ export function useInspectorInfoLonghandShorthand<
       transient?: boolean | undefined,
     ) => {
       const allPropKeysEqual = orderedPropKeys.every((propKeys) => {
-        return arrayEquals(propKeys, orderedPropKeys[0])
+        return arrayEqualsByReference(propKeys, orderedPropKeys[0])
       })
       if (!allPropKeysEqual) {
         // we do nothing for now. we cannot ensure that we can make a sensible update and surface it on the UI as well
@@ -234,7 +234,7 @@ export function useInspectorInfoLonghandShorthand<
   const longhandResultsWithUnset = objectMap((longhandResult, longhandToUnset) => {
     const onUnsetValues = () => {
       const allPropKeysEqual = allOrderedPropKeys.every((propKeys) => {
-        return arrayEquals(propKeys, allOrderedPropKeys[0])
+        return arrayEqualsByReference(propKeys, allOrderedPropKeys[0])
       })
       if (!allPropKeysEqual) {
         // we do nothing for now. we cannot ensure that we can make a sensible update and surface it on the UI as well