--- a/editor/src/core/shared/utils.ts
+++ b/editor/src/core/shared/utils.ts
@@ -30,11 +30,31 @@ export function fastForEach<T>(a: readonly T[], fn: (t: T, index: number) => voi
   }
 }
 
-export function arrayEquals<T>(a: Array<T>, b: Array<T>, eq?: (l: T, r: T) => boolean): boolean {
+export function arrayEqualsByReference<T>(a: Array<T>, b: Array<T>): boolean {
   if (a === b) {
     return true
   } else {
-    const equals = eq == null ? (l: T, r: T) => l === r : eq
+    if (a.length === b.length) {
+      for (let i = 0; i < a.length; i++) {
+        if (a[i] !== b[i]) {
+          return false
+        }
+      }
+      return true
+    } else {
+      return false
+    }
+  }
+}
+
+export function arrayEqualsByValue<T>(
+  a: Array<T>,
+  b: Array<T>,
+  equals: (l: T, r: T) => boolean,
+): boolean {
+  if (a === b) {
+    return true
+  } else {
     if (a.length === b.length) {
       for (let i = 0; i < a.length; i++) {
         if (!equals(a[i], b[i])) {