--- a/editor/src/templates/editor.tsx
+++ b/editor/src/templates/editor.tsx
@@ -14,7 +14,7 @@ import {
   PRODUCTION_ENV,
   requireElectron,
 } from '../common/env-vars'
-import { arrayEquals, EditorID } from '../core/shared/utils'
+import { EditorID } from '../core/shared/utils'
 import CanvasActions from '../components/canvas/canvas-actions'
 import {
   DispatchPriority,