**Problem:**
`PropertyPath` is similar to `ElementPath` in a few ways, so we can probably apply what we've learnt with the latter to the former.

**Fix:**
Improvements were as follows:
- Split apart `arrayEquals` into two versions that have more specific functionality which appears to make it more amenable to optimisations.
- Handle strings and numbers in property paths by having a cache value for each instead of templating a string to use as a key.
- Removed the null handling from `pathsEqual` as we never actually utilised it.

**Results:**
```
creating a property path, same path repeated:
4,317,459 ops/s, ±0.49% -> 89,242,985 ops/s, ±1.26%
creating a property path, new path each time:
672,713 ops/s, ±25.69% -> 2,276,563 ops/s, ±40.46%
creating a property path from an array, same path repeated:
3,486,147 ops/s, ±0.44% -> 22,118,001 ops/s, ±2.17%
creating a property path from an array, new path each time:
951,396 ops/s, ±17.46% -> 1,649,162 ops/s, ±52.37%
toString a property path, same path repeated:
2,494,425 ops/s, ±26.26% -> 32,971,920 ops/s, ±2.34%
toString a property path, new path each time:
1,536,574 ops/s, ±1.36% -> 2,198,471 ops/s, ±4.22%
pathsEqual, same instance:
1,287,738,732 ops/s, ±0.24% -> 1,241,210,081 ops/s, ±0.76%
pathsEqual, same path different instance:
30,391,699 ops/s, ±0.62% -> 180,323,832 ops/s, ±0.79%
pathsEqual, different paths:
32,415,262 ops/s, ±0.55% -> 186,843,290 ops/s, ±2.99%
```

**Commit Details:**
- Split `arrayEquals` into `arrayEqualsByReference` and `arrayEqualsByValue`.
- `PropertyPathCache` now has `stringChildCaches` and `numberChildCaches`, to maintain the difference of indexing by strings or numbers.
- Simplified `pathsEqual` as it doesn't actually need to handle nulls.